requirements.txt
h5py==3.1.0
keras_applications==1.0.8
numpy==1.19.5
pandas==1.3.3
Pillow==8.3.1
Pillow==10.1.0
scikit_learn==1.3.2
streamlit==1.0.0
tensorflow==2.5.0
tensorflow_gpu==2.5.0
tqdm==4.61.2
